feat(features): move Revert to features package
