+++ space-client/src/main/java/io/github/pgmarc/space/features/Revert.java
@@ -1,0 +1,16 @@
+package io.github.pgmarc.space.features;
+
+public enum Revert {
+    OLDEST_VALUE(false),
+    NEWEST_VALUE(true);
+
+    private final boolean latest;
+
+    Revert(boolean latest) {
+        this.latest = latest;
+    }
+
+    public boolean isLatest() {
+        return latest;
+    }
+}
